Refactor: Fixed indexOf() & isDuplicate() method logic.
Added a check for current Node data being null, & data being null.

if current data in list is null, then if given data is null, return index, instead of trying to compare two null objects or an object to a null. Also, small changes to spacing in rest of file.